requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.38.0
django-environ==0.4.5
django-filter==2.2.0
django-markdownx==3.0.1
django-rest-auth==0.9.3
django-tinymce==2.8.0
djangorestframework==3.11.0
gunicorn==19.10.0
idna==2.8
Markdown==3.2.1
oauthlib==2.0.3
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==0.8.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.8

